Handle chown with no user or group (#10637)

+++ pkg/security/secl/model/accessors.go
@@ -14219,7 +14219,7 @@ func (e *Event) SetFieldValue(field eval.Field, value interface{}) error {
 		if !ok {
 			return &eval.ErrValueTypeMismatch{Field: "Chown.GID"}
 		}
-		e.Chown.GID = uint32(v)
+		e.Chown.GID = int64(v)
 
 		return nil
 
@@ -14241,7 +14241,7 @@ func (e *Event) SetFieldValue(field eval.Field, value interface{}) error {
 		if !ok {
 			return &eval.ErrValueTypeMismatch{Field: "Chown.UID"}
 		}
-		e.Chown.UID = uint32(v)
+		e.Chown.UID = int64(v)
 
 		return nil
